Move compile entrypoint
